# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for 偶然の6人
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: number;
  clerkUserId: string;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  department?: string | null;
  role: string;
  interestTags: string[];
  profileComplete: boolean;
  createdAt: string;
}

export interface UserProfileUpdate {
  name?: string;
  department?: string;
  interestTags?: string[];
  role?: string;
}

export interface UserStats {
  participated: number;
  hosted: number;
  connections: number;
}

export interface Connection {
  id: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  department?: string | null;
  metAtEventId: number;
  metAtEventTheme: string;
}

export type EventStatus = typeof EventStatus[keyof typeof EventStatus];


export const EventStatus = {
  募集中: '募集中',
  実施確定: '実施確定',
  未実施: '未実施',
  開催済: '開催済',
} as const;

export interface Event {
  id: number;
  theme: string;
  /** @nullable */
  subTheme?: string | null;
  datetime: string;
  durationMinutes: number;
  location: string;
  /** @nullable */
  locationUrl?: string | null;
  fee: number;
  capacity: number;
  minParticipants: number;
  deadline: string;
  /** @nullable */
  notes?: string | null;
  tags: string[];
  hostId: number;
  /** @nullable */
  hostName?: string | null;
  /** @nullable */
  hostDepartment?: string | null;
  participantsCount: number;
  remainingSeats: number;
  status: EventStatus;
  isApplied: boolean;
  isHost: boolean;
  createdAt: string;
}

export interface EventInput {
  /** @minLength 1 */
  theme: string;
  subTheme?: string;
  datetime: string;
  /**
     * @minimum 1
     * @maximum 120
     */
  durationMinutes: number;
  /** @minLength 1 */
  location: string;
  locationUrl?: string;
  /**
     * @minimum 0
     * @maximum 5000
     */
  fee: number;
  /**
     * @minimum 3
     * @maximum 6
     */
  capacity: number;
  /**
     * @minimum 2
     * @maximum 6
     */
  minParticipants: number;
  deadline: string;
  notes?: string;
  tags: string[];
}

export interface EventUpdate {
  theme?: string;
  subTheme?: string;
  datetime?: string;
  durationMinutes?: number;
  location?: string;
  locationUrl?: string;
  fee?: number;
  capacity?: number;
  minParticipants?: number;
  deadline?: string;
  notes?: string;
  tags?: string[];
}

export interface Participation {
  id: number;
  eventId: number;
  userId: number;
  status: string;
  /** @nullable */
  comment?: string | null;
  appliedAt: string;
}

export interface ParticipantInfo {
  userId: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  comment?: string | null;
  appliedAt: string;
}

export interface ApplicationWithEvent {
  id: number;
  eventId: number;
  status: string;
  /** @nullable */
  comment?: string | null;
  appliedAt: string;
  event?: Event | null;
}

export type ReportType = typeof ReportType[keyof typeof ReportType];


export const ReportType = {
  参加者感想: '参加者感想',
  開催者報告: '開催者報告',
} as const;

export interface Report {
  id: number;
  eventId: number;
  authorId: number;
  /** @nullable */
  authorName?: string | null;
  type: ReportType;
  content: string;
  /** @nullable */
  photoUrl?: string | null;
  likesCount: number;
  isLiked: boolean;
  createdAt: string;
}

export type ReportInputType = typeof ReportInputType[keyof typeof ReportInputType];


export const ReportInputType = {
  参加者感想: '参加者感想',
  開催者報告: '開催者報告',
} as const;

export interface ReportInput {
  type: ReportInputType;
  /** @minLength 1 */
  content: string;
  photoUrl?: string;
}

export interface LikeResult {
  liked: boolean;
  likesCount: number;
}

export interface Notification {
  id: number;
  type: string;
  content: string;
  isRead: boolean;
  createdAt: string;
}

export interface FeedbackInput {
  /** @minLength 1 */
  content: string;
  eventId?: number;
}

export interface FeedbackItem {
  id: number;
  content: string;
  /** @nullable */
  fromUserName?: string | null;
  fromUserEmail: string;
  /** @nullable */
  eventId?: number | null;
  /** @nullable */
  eventTheme?: string | null;
  createdAt: string;
}

export type OfficeKpiTopTagsItem = {
  tag: string;
  count: number;
};

export interface OfficeKpi {
  totalEvents: number;
  completedEvents: number;
  cancelledEvents: number;
  eventsThisMonth: number;
  totalParticipants: number;
  participantsThisMonth: number;
  totalConnectionPairs: number;
  totalFeedbacks: number;
  topTags: OfficeKpiTopTagsItem[];
}

export type ListEventsParams = {
status?: string;
tag?: string;
sortBy?: ListEventsSortBy;
};

export type ListEventsSortBy = typeof ListEventsSortBy[keyof typeof ListEventsSortBy];


export const ListEventsSortBy = {
  new: 'new',
  near: 'near',
  seats: 'seats',
} as const;

export type ApplyToEventBody = {
  comment?: string;
};