# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: API specification for 偶然の6人
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [new, near, seats]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update event (host only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event (host only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
...
  /my/applications:
    get:
      operationId: listMyApplications
      tags: [participations]
      summary: My participations
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicationWithEvent"
        "401":
          description: Unauthorized
  /my/hosted-events:
    get:
      operationId: listMyHostedEvents
      tags: [events]
      summary: My hosted events
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
        "401":
          description: Unauthorized
  /my/stats:
    get:
      operationId: getMyStats
      tags: [users]
      summary: My activity stats
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"
        "401":
          description: Unauthorized
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: My notifications
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
        "401":
          description: Unauthorized
  /notifications/read-all:
    patch:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications read
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
  /feedbacks:
    post:
      operationId: submitFeedback
      tags: [feedbacks]
      summary: Submit feedback to office
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackInput"
      responses:
        "201":
          description: Submitted
        "401":
          description: Unauthorized
...
  /office/feedbacks:
    get:
      operationId: listOfficeFeedbacks
      tags: [office]
      summary: List feedbacks (office only)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeedbackItem"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /office/export/events:
    get:
      operationId: exportEventsCSV
      tags: [office]
      summary: Export events as CSV (office only)
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
components:
  schemas:
...
    UserStats:
      type: object
      properties:
        participated:
          type: integer
        hosted:
          type: integer
        connections:
          type: integer
      required: [participated, hosted, connections]
...
    Event:
      type: object
      properties:
        id:
          type: integer
        theme:
          type: string
        subTheme:
          type: ["string", "null"]
        datetime:
          type: string
        durationMinutes:
          type: integer
        location:
          type: string
        locationUrl:
          type: ["string", "null"]
        fee:
          type: integer
        capacity:
          type: integer
        minParticipants:
          type: integer
        deadline:
          type: string
        notes:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        hostId:
          type: integer
        hostName:
          type: ["string", "null"]
        hostDepartment:
          type: ["string", "null"]
        participantsCount:
          type: integer
        remainingSeats:
          type: integer
        status:
          type: string
          enum: [募集中, 実施確定, 未実施, 開催済]
        isApplied:
          type: boolean
        isHost:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - theme
        - datetime
        - durationMinutes
        - location
        - fee
        - capacity
        - minParticipants
        - deadline
        - tags
        - hostId
        - participantsCount
        - remainingSeats
        - status
        - isApplied
        - isHost
        - createdAt
    EventInput:
      type: object
      properties:
        theme:
          type: string
          minLength: 1
        subTheme:
          type: string
        datetime:
          type: string
        durationMinutes:
          type: integer
          minimum: 1
          maximum: 120
        location:
          type: string
          minLength: 1
        locationUrl:
          type: string
        fee:
          type: integer
          minimum: 0
          maximum: 5000
        capacity:
          type: integer
          minimum: 3
          maximum: 6
        minParticipants:
          type: integer
          minimum: 2
          maximum: 6
        deadline:
          type: string
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - theme
        - datetime
        - durationMinutes
        - location
        - fee
        - capacity
        - minParticipants
        - deadline
        - tags
    EventUpdate:
      type: object
      properties:
        theme:
          type: string
        subTheme:
          type: string
        datetime:
          type: string
        durationMinutes:
          type: integer
        location:
          type: string
        locationUrl:
          type: string
        fee:
          type: integer
        capacity:
          type: integer
        minParticipants:
          type: integer
        deadline:
          type: string
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
...
    ApplicationWithEvent:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        status:
          type: string
        comment:
          type: ["string", "null"]
        appliedAt:
          type: string
        event:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Event"
      required: [id, eventId, status, appliedAt]
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        content:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required: [id, type, content, isRead, createdAt]
    FeedbackInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
        eventId:
          type: integer
      required: [content]
    FeedbackItem:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        fromUserName:
          type: ["string", "null"]
        fromUserEmail:
          type: string
        eventId:
          type: ["integer", "null"]
        eventTheme:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, content, fromUserEmail, createdAt]
...